requirements.txt
annotated-types==0.7.0
anyio==4.7.0
bcrypt==3.2.2
certifi==2024.12.14
cffi==1.17.1
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.8.0
pytest==8.0.2
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.12.2
uvicorn==0.29.0
